create a separate errors in packagelock.json
